Use single health.lua for all checks
